fixed issue where multiple submissions from one student would mess up others' naming

+++ mossScript.py
@@ -3,6 +3,7 @@ import os
 import shutil
 import subprocess
 import platform
+import random
 
 # Deciding on directory sep. Mac&Linux us '/' while Windows uses '\'
 os_separator: str = "/" if platform.system() in ["Linux", "Darwin"] else "\\"
@@ -23,15 +24,18 @@ curr_dir = os.path.abspath('.') + os_separator
 test_exist = curr_dir+os_separator+'StudentCode'+os_separator+'namedFiles'
 if os.path.exists(test_exist) and os.listdir(test_exist):
     y_n = str(input(
-        "namedFiles already exists and contains files. Would you like to empty it?")).lower()[0]
-    if y_n == 'y':
+        "namedFiles already exists and contains files. Would you like to empty it?")).lower()
+    if len(y_n) == 0 or y_n[0] == 'y':
         shutil.rmtree(test_exist)
 
 # Choosing a file extension
 java_or_else = str(
-    input('Use .java file extension? if not, enter the file extension to use (.cpp,.py,etc)'))
+    input('Use .java file extension? if not, enter the file extension to use (.cpp,.py,etc)\n'))
 file_extension = '.java' if len(java_or_else) == 0 or java_or_else.lower()[
     0] == 'y' else java_or_else
+# add . if not present
+file_extension = '.' + \
+    file_extension if file_extension[0] != '.' else file_extension
 print('Using {} as file extension and {} as directory seperator'.format(
     file_extension, os_separator))
 
@@ -55,10 +59,13 @@ def unzip():
         shutil.rmtree(unzipped_dir)
     except:
         pass
-    file_name = [file for file in os.listdir(dir) if file[-4:] == '.zip'][0]
+
+    file_name = [file for file in os.listdir(dir) if file[-4:] == '.zip']
+
+    file_name = file_name[0] if len(file_name) > 0 else None
     if not file_name:
-        print('no zip files found')
-        exit(0)
+        print('no additional zip files found')
+        return
     with ZipFile(dir+os_separator+file_name, 'r') as zip:
         os.mkdir(unzipped_dir)
         zip.extractall(unzipped_dir)
@@ -71,9 +78,17 @@ def file_to_dir():
     """
     if not os.path.exists(dir+os_separator+'namedFiles'):
         os.mkdir(dir+os_separator+'namedFiles')
+    if not os.path.exists(unzipped_dir):
+        return
     for file_name in os.listdir(unzipped_dir):
-        if file_name[-5:] == file_extension:
-            new_name = file_name.split(' - ')[1].split()[1] + file_extension
+
+        if file_name[-len(file_extension):] == file_extension:
+            new_name = ''
+            try:
+                new_name = file_name.split(
+                    ' - ')[1].split()[1] + file_extension if file_name.index(' - ') else file_name
+            except:
+                new_name = file_name
             os.rename(unzipped_dir+os_separator+file_name,
                       curr_dir+dir+os_separator+'namedFiles'+os_separator+new_name)
 
@@ -84,12 +99,20 @@ def dealWithZip():
     creates temp lastName.dir folders that the .zip gets extracted to
     """
     i = 0
+    if not os.path.exists(unzipped_dir):
+        return
     for zipped_file in os.listdir(unzipped_dir):
         if zipped_file[-4:] != '.zip':
             continue
         with ZipFile(unzipped_dir+os_separator+zipped_file, 'r') as zf:
+            get_name = ''
+            try:
+                get_name = zipped_file.split(
+                    ' - ')[1].split()[1] if zipped_file.index(' - ') else zipped_file
+            except:
+                get_name = zipped_file
             temp_dir = unzipped_dir+os_separator + \
-                (zipped_file.split(' - ')[1].split()[1])+'.dir'
+                (get_name)+'.dir'
             if not os.path.exists(temp_dir):
                 os.mkdir(temp_dir)
             zf.extractall(temp_dir)
@@ -104,18 +127,23 @@ def find_file_extension_files(search_path):
     """
     results = []
     titles = []  # names of students
-    title = ''
+    title = 'dummyName'
     # Wlaking top-down from the root
+    i = 1
     for root, dir, files in os.walk(search_path):
-        java_files = []
+        code_files = []
         # title = parent folder name. just so we save the name of the student that we are currently in
         title = root.split(
             os_separator)[-1] if root.split(os_separator)[-1][-4:] == '.dir' else title
-        java_files = [os.path.join(root, file)
-                      for file in files if file[-5:] == file_extension and file[0:2] != '._']  # '._' is a MacOS thing
-        if java_files:
-            titles.append(title)
-            results.extend(java_files)
+        i += 1
+        code_files = [os.path.join(root, file)
+                      for file in files if file[-len(file_extension):] == file_extension and file[0:2] != '._']  # '._' is a MacOS thing
+        if code_files:
+            for i in range(len(code_files)):
+                titles.append(title)
+            results.extend(code_files)
+    titles = list(map(lambda x: x+str(random.choice(range(0, 100)))
+                      if x == 'dummyName' else x, titles))
     return results, titles
 
 # moves files in list
@@ -136,6 +164,8 @@ def moveFiles(lst, titles):
                 0]+silly_number+file_extension
             silly_number = str(int(silly_number)+1)
         os.rename(file_path, new_path)
+    if not os.path.exists(unzipped_dir):
+        return
     shutil.rmtree(unzipped_dir)
 
 
@@ -148,15 +178,23 @@ dealWithZip()
 
 # get all .{file_extension} file abs paths
 zip_dirs, titles = find_file_extension_files(unzipped_dir)
-titles = list(map(lambda x: x[0:len(x)-4]+file_extension, titles))
-print('Found {} submissions.'.format(len(titles) + len(os.listdir(named_dir))))
+i = 0
+# Fill missing titles using zip_dirs
+while len(zip_dirs) > len(titles) and len(titles)+i < len(titles)-1:
+    titles.append(zip_dirs[len(titles)+i].split('/')[3])
+    i += 1
+
+titles = list(
+    map(lambda x: (x[0:len(x)-len(file_extension)+1] if len(x) > len(file_extension) else x)+file_extension, titles))
+print('Found {} {} and {} .zip submissions.'.format(
+    len(os.listdir(named_dir)), file_extension, len(zip_dirs)))
 # move .{file_extension} files into namedFiles
 moveFiles(
     zip_dirs, titles)
 
 # # run moss!
 subprocess.Popen(
-    f'.{os_separator}moss .{os_separator}StudentCode{os_separator}namedFiles{os_separator}*.java', shell=True)
+    f'.{os_separator}moss .{os_separator}StudentCode{os_separator}namedFiles{os_separator}*{file_extension}', shell=True)
 
 # WIP
 # mossum = input('run mossum too?').lower()[0] == 'y'
